Fix part 1 bug, test now pass
The problem was that the "address" of the client was added, even though
it was empty.

It is important to note that the convention used in part 1, for a
SimpleMessage, is that a client does not specify any address in the data
structure. The field is simply an empty string.